Distinguish between component creation in init and via DI command
Previously, we added an `isInitialization` flag in all the
`add[Component]` functions that could be called both from
`createComponentAndView` and from the toolbar.

The assumption was that `createComponentAndView` was only called
during initialization (when loading a document with components),
so `createComponentAndView` always passed down `isInitialization=true`.

Now `createComponentAndView` can also be called via a `createComponent`
call from the data interactive.

To distinguish between the two, we add an additional argument to the
component arguments, `initiatedViaCommand`.

However, there may be other ways to distinguish between the two events.